# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Standup Comedy Connect API",
    "description": "API pour la plateforme de mise en relation entre humoristes et organisateurs d'événements de stand-up comedy",
    "version": "1.0.0",
    "contact": {
      "name": "Support API",
      "email": "support@standup-comedy-connect.com"
    }
  },
  "servers": [
    {
      "url": "https://standup-comedy-app.onrender.com/api",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Serveur de développement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/absences": {
      "post": {
        "tags": ["Absences"],
        "summary": "Marquer un participant comme absent",
        "description": "Permet à un organisateur de marquer un humoriste comme absent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId", "comedianId"],
                "properties": {
                  "eventId": {"type": "string"},
                  "comedianId": {"type": "string"},
                  "reason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Absence marquée avec succès"
          }
        }
      }
    },
    "/absences/event/{eventId}": {
      "get": {
        "tags": ["Absences"],
        "summary": "Récupérer les absences d'un événement",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des absences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {"type": "string"},
                      "comedian": {"type": "object"},
                      "organizer": {"type": "object"},
                      "reason": {"type": "string"},
                      "markedAt": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}